CHG: refine the library path; add __init__ for sssd; add a new tests folder
