+++ sssd/imputers/DiffWaveImputer.py
@@ -2,7 +2,8 @@ import math
 
 import torch
 import torch.nn as nn
-from utils.util import calc_diffusion_step_embedding
+
+from sssd.utils.util import calc_diffusion_step_embedding
 
 
 def swish(x):
