+++ sssd/imputers/SSSDSAImputer.py
@@ -2,8 +2,9 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from einops import rearrange
-from imputers.S4Model import S4, LinearActivation
-from utils.util import calc_diffusion_step_embedding
+
+from sssd.imputers.S4Model import S4, LinearActivation
+from sssd.utils.util import calc_diffusion_step_embedding
 
 
 def swish(x):
