--- sssd/inference_multiples.py
@@ -1,275 +1,0 @@
-import argparse
-import datetime
-import json
-import os
-from statistics import mean
-
-import numpy as np
-import torch
-import torch.nn as nn
-from imputers.DiffWaveImputer import DiffWaveImputer
-from imputers.SSSDS4Imputer import SSSDS4Imputer
-from imputers.SSSDSAImputer import SSSDSAImputer
-from sklearn.metrics import mean_squared_error
-from utils.util import (
-    calc_diffusion_hyperparams,
-    find_max_epoch,
-    get_mask_bm,
-    get_mask_forecast,
-    get_mask_mnr,
-    get_mask_rm,
-    print_size,
-    sampling,
-)
-
-
-def generate(
-    output_directory,
-    num_samples,
-    ckpt_path,
-    data_path,
-    ckpt_iter,
-    use_model,
-    masking,
-    missing_k,
-    only_generate_missing,
-    index,
-):
-
-    """
-    Generate data based on ground truth
-
-    Parameters:
-    output_directory (str):           save generated speeches to this path
-    num_samples (int):                number of samples to generate, default is 4
-    ckpt_path (str):                  checkpoint path
-    ckpt_iter (int or 'max'):         the pretrained checkpoint to be loaded;
-                                      automitically selects the maximum iteration if 'max' is selected
-    data_path (str):                  path to NYISO, numpy array.
-    use_model (int):                  0:DiffWave. 1:SSSDSA. 2:SSSDS4.
-    masking (str):                    'mnr': missing not at random, 'bm': black-out, 'rm': random missing
-    only_generate_missing (int):      0:all sample diffusion.  1:only apply diffusion to missing portions of the signal
-    missing_k (int)                   k missing time points for each channel across the length.
-    """
-
-    # generate experiment (local) path
-    local_path = "T{}_beta0{}_betaT{}".format(
-        diffusion_config["T"], diffusion_config["beta_0"], diffusion_config["beta_T"]
-    )
-
-    # map diffusion hyperparameters to gpu
-    for key in diffusion_hyperparams:
-        if key != "T":
-            diffusion_hyperparams[key] = diffusion_hyperparams[key].cuda()
-
-    # predefine model
-    if use_model == 0:
-        net = DiffWaveImputer(**model_config).cuda()
-    elif use_model == 1:
-        net = SSSDSAImputer(**model_config).cuda()
-    elif use_model == 2:
-        net = SSSDS4Imputer(**model_config).cuda()
-    else:
-        print("Model chosen not available.")
-    print_size(net)
-
-    # Check if multiple GPUs are available
-    if torch.cuda.device_count() > 1:
-        print("Using ", torch.cuda.device_count(), " GPUs!")
-        net = nn.DataParallel(net)
-
-    # Move the model to the GPU(s)
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    net.to(device)
-
-    # load checkpoint
-    ckpt_path = os.path.join(ckpt_path, local_path)
-    if ckpt_iter == "max":
-        ckpt_iter = find_max_epoch(ckpt_path)
-    model_path = os.path.join(ckpt_path, "{}.pkl".format(ckpt_iter))
-
-    try:
-        checkpoint = torch.load(model_path, map_location="cpu")
-        print("checkpoint")
-        net.load_state_dict(checkpoint["model_state_dict"])
-        print("Successfully loaded model at iteration {}".format(ckpt_iter))
-    except:
-        raise Exception("No valid model found")
-
-    # Get shared output_directory ready
-    output_directory = os.path.join(
-        output_directory,
-        local_path,
-        "imputaiton_multiple_" + str(round(int(ckpt_iter) / 1000)) + "k",
-    )
-    if not os.path.isdir(output_directory):
-        os.makedirs(output_directory)
-        os.chmod(output_directory, 0o775)
-    print("output directory", output_directory, flush=True)
-
-    ### Custom data loading and reshaping ###
-
-    testing_data = np.load(trainset_config["test_data_path"])
-    ####################################
-    # testing_data = testing_data[0:4418,]
-    # only for TN, 現在 4423 無法被整除，用回原本的 4418
-    # testing_data.shape[0] < 4418 時，這行指令不會對 testing_data 有影響
-    ####################################
-    testing_data = np.split(
-        testing_data, testing_data.shape[0] / num_samples, 0
-    )  # (data, 分4組, 0) batch size = num_samples = 272 為了創造 batch，除不盡可用 np.array_split
-    testing_data = np.array(testing_data)
-    testing_data = torch.from_numpy(testing_data).float().cuda()
-    print("Data loaded")
-
-    all_mse = []
-
-    for i, batch in enumerate(testing_data):
-
-        if masking == "mnr":
-            mask_T = get_mask_mnr(batch[0], missing_k)
-            mask = mask_T.permute(1, 0)
-            mask = mask.repeat(batch.size()[0], 1, 1)
-            mask = mask.type(torch.float).cuda()
-
-        elif masking == "bm":
-            mask_T = get_mask_bm(batch[0], missing_k)
-            mask = mask_T.permute(1, 0)
-            mask = mask.repeat(batch.size()[0], 1, 1)
-            mask = mask.type(torch.float).cuda()
-
-        elif masking == "rm":
-            mask_T = get_mask_rm(batch[0], missing_k)
-            mask = mask_T.permute(1, 0)
-            mask = mask.repeat(batch.size()[0], 1, 1).float().cuda()
-
-        elif masking == "forecast":
-            mask_T = get_mask_forecast(batch[0], missing_k)
-            mask = mask_T.permute(1, 0)
-            mask = mask.repeat(batch.size()[0], 1, 1)
-            mask = mask.type(torch.float).cuda()
-
-        batch = batch.permute(0, 2, 1)
-
-        start = torch.cuda.Event(enable_timing=True)
-        end = torch.cuda.Event(enable_timing=True)
-        start.record()
-
-        sample_length = batch.size(2)
-        sample_channels = batch.size(1)
-        generated_audio = sampling(
-            net,
-            (num_samples, sample_channels, sample_length),
-            diffusion_hyperparams,
-            cond=batch,
-            mask=mask,
-            only_generate_missing=only_generate_missing,
-        )
-
-        end.record()
-        torch.cuda.synchronize()
-
-        print(
-            "generated {} utterances of random_digit at iteration {} in {} seconds".format(
-                num_samples, ckpt_iter, int(start.elapsed_time(end) / 1000)
-            )
-        )
-
-        generated_audio = generated_audio.detach().cpu().numpy()
-        batch = batch.detach().cpu().numpy()
-        mask = mask.detach().cpu().numpy()
-
-        outfile = f"{index}_imputation{i}.npy"
-        new_out = os.path.join(output_directory, outfile)
-        np.save(new_out, generated_audio)
-
-        outfile = f"original{i}.npy"
-        new_out = os.path.join(output_directory, outfile)
-        np.save(new_out, batch)
-
-        outfile = f"mask{i}.npy"
-        new_out = os.path.join(output_directory, outfile)
-        np.save(new_out, mask)
-
-        print("saved generated samples at iteration %s" % ckpt_iter)
-
-        mse = mean_squared_error(
-            generated_audio[~mask.astype(bool)], batch[~mask.astype(bool)]
-        )
-        all_mse.append(mse)
-
-    print("Total MSE:", mean(all_mse))
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "-c",
-        "--config",
-        type=str,
-        default="config.json",
-        help="JSON file for configuration",
-    )
-    parser.add_argument(
-        "-n",
-        "--num_samples",
-        type=int,
-        default=1,
-        help="Number of utterances to be generated",
-    )
-    parser.add_argument(
-        "-ckpt_iter",
-        "--ckpt_iter",
-        default="max",
-        help='Which checkpoint to use; assign a number or "max"',
-    )
-    args = parser.parse_args()
-
-    # Parse configs. Globals nicer in this case
-    with open(args.config) as f:
-        data = f.read()
-    config = json.loads(data)
-    print(config)
-
-    gen_config = config["gen_config"]
-
-    train_config = config["train_config"]  # training parameters
-
-    global trainset_config
-    trainset_config = config["trainset_config"]  # to load trainset
-
-    global diffusion_config
-    diffusion_config = config["diffusion_config"]  # basic hyperparameters
-
-    global diffusion_hyperparams
-    diffusion_hyperparams = calc_diffusion_hyperparams(
-        **diffusion_config
-    )  # dictionary of all diffusion hyperparameters
-
-    global model_config
-    if train_config["use_model"] == 0:
-        model_config = config["wavenet_config"]
-    elif train_config["use_model"] == 1:
-        model_config = config["sashimi_config"]
-    elif train_config["use_model"] == 2:
-        model_config = config["wavenet_config"]
-
-    for i in range(10):
-        print(i)
-        current_time = datetime.datetime.now()
-        print("當前時間:", current_time.strftime("%Y-%m-%d %H:%M:%S"))
-
-        generate(
-            **gen_config,
-            ckpt_iter=args.ckpt_iter,
-            num_samples=args.num_samples,
-            use_model=train_config["use_model"],
-            data_path=trainset_config["test_data_path"],
-            masking=train_config["masking"],
-            missing_k=train_config["missing_k"],
-            only_generate_missing=train_config["only_generate_missing"],
-            index=i,
-        )
-
-    current_time = datetime.datetime.now()
-    print("當前時間:", current_time.strftime("%Y-%m-%d %H:%M:%S"))
